Consolidate Crypto module into namespace enum with Data extensions
