--- YubiKit/YubiKit/Crypto/CryptoKit+Extensions.swift
@@ -1,25 +1,0 @@
-// Copyright Yubico AB
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import CryptoKit
-import Foundation
-
-extension CryptoKit.Digest {
-    fileprivate var bytes: [UInt8] { Array(makeIterator()) }
-    fileprivate var data: Data { Data(bytes) }
-
-    fileprivate var hexStr: String {
-        bytes.map { String(format: "%02X", $0) }.joined()
-    }
-}
